Auto-fix compiler warnings re. NSUInteger in formatted strings

+++ NBClient/NBClient/NBClient+People.m
@@ -41,7 +41,7 @@
 {
     NSURLComponents *components = self.baseURLComponents.copy;
     components.path = [components.path stringByAppendingString:
-                       [NSString stringWithFormat:@"/people/%d", identifier]];
+                       [NSString stringWithFormat:@"/people/%lu", (unsigned long)identifier]];
     return [self baseFetchTaskWithURLComponents:components resultsKey:@"person" completionHandler:completionHandler];
 }
 
@@ -79,7 +79,7 @@
 {
     NSURLComponents *components = self.baseURLComponents.copy;
     components.path = [components.path stringByAppendingString:
-                       [NSString stringWithFormat:@"/people/%d", identifier]];
+                       [NSString stringWithFormat:@"/people/%lu", (unsigned long)identifier]];
     NSError *error;
     NSMutableURLRequest *request = [self baseSaveRequestWithURL:components.URL parameters:@{ @"person": parameters } error:&error];
     if (error) {
@@ -94,7 +94,7 @@
 {
     NSURLComponents *components = self.baseURLComponents.copy;
     components.path = [components.path stringByAppendingString:
-                       [NSString stringWithFormat:@"/people/%d", identifier]];
+                       [NSString stringWithFormat:@"/people/%lu", (unsigned long)identifier]];
     return [self baseDeleteTaskWithURL:components.URL completionHandler:completionHandler];
 }
